requirements.txt
Django>=4.2.0
django-allauth>=0.54.0
django-crispy-forms>=2.0
django-environ>=0.10.0
pillow>=10.0.0
python-dateutil>=2.8.2
pytz>=2023.3
whitenoise>=6.5.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.6
django-widget-tweaks>=1.4.12
dj-database-url>=2.1.0
python-dotenv>=1.0.0
